refactor(dev-workflow/feed-platform-ms-01-shared-libraries): restore barrel exports per original design
User feedback (2026-05-11): fp などの library barrel export は許容方針で oxlint-disable
コメント (oxlint-disable-next-line oxc/no-barrel-file -- ...) で運用しているため、
当初 design.md L210-216 で想定していた barrel パターンに戻す。modern bundler の
解析能力向上により従来の "no-barrel-file 100 module threshold" 制約は緩和できる前提。

実装変更:
- js/package/effect-hono/src/index.ts (新規、export * as Env + dynamicLoggerLayer + makeDisposableRuntime の barrel)
- js/package/remix-helper/src/index.ts (新規、export * from './frame-helpers.ts' の barrel)
- 2 library package.json を subpath exports (./env / ./logger / ./runtime, ./frame-helpers) から単一 "." export に統一
- 各 export * 行に // oxlint-disable-next-line oxc/no-barrel-file 注釈 (fp/src/effect.ts 同形)

consumer 側 import の統合:
- 3 consumer (backend / web / IdP) の feature/env.ts 再エクスポート shim を削除 (旧 effect-hono/env subpath 経由)
- health.ts / health.test.ts: `import * as Env from './env.ts'` → `import { Env } from 'effect-hono'`
- runtime/server.ts: 3 つの subpath import (effect-hono/logger + effect-hono/runtime + '../env.ts')
  → 1 barrel import `import { dynamicLoggerLayer, Env, makeDisposableRuntime } from 'effect-hono'`

検証:
- vp run check exit 0 (formatting / lint / typecheck 全 PASS)
- TC-001 effect-hono runtime smoke 2/2 PASS
- TC-002 remix-helper frame-helpers smoke 2/2 PASS
- backend health.test.ts 1/1 PASS (regression なし)
- SC-5 旧コード残存 0 hit (consumer 内の再 export / 関数呼び出し起源の hit のみ、旧 inline 定義は不在)
- SC-6 createFrameHelpers 利用: web 2 / IdP 2 / hono-example 4 hit ≥ 1
