--- js/app/feed-platform-backend/src/feature/env.ts
@@ -1,2 +1,0 @@
-export { makeLayer, Service, layer } from 'effect-hono/env'
-export type { Type } from 'effect-hono/env'
