--- js/app/feed-platform-web/app/feature/env.ts
@@ -1,2 +1,0 @@
-export { makeLayer, Service, layer } from 'effect-hono/env'
-export type { Type } from 'effect-hono/env'
