+++ js/package/remix-helper/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "exports": {
-    ".": "./src/frame-helpers.ts"
+    ".": "./src/index.ts"
   },
   "devDependencies": {
     "@totto2727/fp": "workspace:*",
